Database for username and password login enabled
The app now accepts a user name and password and searches the database
for them. Currently the only username and passwords are "user" and
"pass". Looking to extend these updates to embedding the user name in
the dashboard view. For now I am planning on using cookies, will find
another work around soon.